refactor: renaming stuff moving func
