+++ src/table/components/TablePaginationActions.jsx
@@ -6,7 +6,7 @@ import KeyboardArrowLeft from '@material-ui/icons/KeyboardArrowLeft';
 import KeyboardArrowRight from '@material-ui/icons/KeyboardArrowRight';
 import LastPageIcon from '@material-ui/icons/LastPage';
 import { makeStyles } from '@material-ui/core/styles';
-import { focusConfirmButton } from '../utils/handle-cell-focus';
+import { focusConfirmButton } from '../utils/handle-accessibility';
 
 const useStyles = makeStyles({
   root: {
