+++ src/table/utils/__tests__/handle-accessiblity.spec.js
@@ -1,7 +1,7 @@
 import { expect } from 'chai';
-import * as handleCellFocus from '../handle-cell-focus';
+import * as handleAccessibility from '../handle-accessibility';
 
-describe('handle-key-press', () => {
+describe('handle-accessibility', () => {
   let cell;
   let rootElement;
   let focusedCellCoord;
@@ -39,7 +39,7 @@ describe('handle-key-press', () => {
     });
 
     it('should focus cell and call setAttribute when focusType is focus', () => {
-      handleCellFocus.updateFocus({ focusType, rowElements, cellCoord });
+      handleAccessibility.updateFocus({ focusType, rowElements, cellCoord });
       expect(cell.focus).to.have.been.calledOnce;
       expect(cell.setAttribute).to.have.been.calledOnceWith('tabIndex', '0');
     });
@@ -47,7 +47,7 @@ describe('handle-key-press', () => {
     it('should blur cell and call setAttribute when focusType is blur', () => {
       focusType = 'blur';
 
-      handleCellFocus.updateFocus({ focusType, rowElements, cellCoord });
+      handleAccessibility.updateFocus({ focusType, rowElements, cellCoord });
       expect(cell.blur).to.have.been.calledOnce;
       expect(cell.setAttribute).to.have.been.calledOnceWith('tabIndex', '-1');
     });
@@ -55,7 +55,7 @@ describe('handle-key-press', () => {
     it('should call setAttribute when focusType is addTab', () => {
       focusType = 'addTab';
 
-      handleCellFocus.updateFocus({ focusType, rowElements, cellCoord });
+      handleAccessibility.updateFocus({ focusType, rowElements, cellCoord });
       expect(cell.focus).to.not.have.been.called;
       expect(cell.setAttribute).to.have.been.calledOnceWith('tabIndex', '0');
     });
@@ -63,7 +63,7 @@ describe('handle-key-press', () => {
     it('should call setAttribute when focusType is removeTab', () => {
       focusType = 'removeTab';
 
-      handleCellFocus.updateFocus({ focusType, rowElements, cellCoord });
+      handleAccessibility.updateFocus({ focusType, rowElements, cellCoord });
       expect(cell.blur).to.not.have.been.called;
       expect(cell.setAttribute).to.have.been.calledOnceWith('tabIndex', '-1');
     });
@@ -71,7 +71,7 @@ describe('handle-key-press', () => {
     it('should not focus cell nor cell setAttribute when cell is not found', () => {
       cellCoord = [1, 0];
 
-      handleCellFocus.updateFocus({ focusType, rowElements, cellCoord });
+      handleAccessibility.updateFocus({ focusType, rowElements, cellCoord });
       expect(cell.focus).to.not.have.been.called;
       expect(cell.blur).to.not.have.been.called;
       expect(cell.setAttribute).to.not.have.been.called;
@@ -80,7 +80,7 @@ describe('handle-key-press', () => {
     it('should pick up providedCell element if there was no rowElements provided', () => {
       providedCell = cell;
 
-      handleCellFocus.updateFocus({ focusType, providedCell });
+      handleAccessibility.updateFocus({ focusType, providedCell });
       expect(cell.focus).to.have.been.calledOnce;
       expect(cell.setAttribute).to.have.been.calledOnceWith('tabIndex', '0');
     });
@@ -105,7 +105,7 @@ describe('handle-key-press', () => {
     it('should return active td element', () => {
       cell = elementCreator('td', '0');
 
-      const cellElement = handleCellFocus.findCellWithTabStop(rootElement);
+      const cellElement = handleAccessibility.findCellWithTabStop(rootElement);
 
       expect(cellElement).to.not.be.null;
       expect(cellElement.tagName).to.equal('TD');
@@ -115,7 +115,7 @@ describe('handle-key-press', () => {
     it('should return active th element', () => {
       cell = elementCreator('th', '0');
 
-      const cellElement = handleCellFocus.findCellWithTabStop(rootElement);
+      const cellElement = handleAccessibility.findCellWithTabStop(rootElement);
 
       expect(cellElement).to.not.be.null;
       expect(cellElement.tagName).to.equal('TH');
@@ -125,7 +125,7 @@ describe('handle-key-press', () => {
     it('should return null', () => {
       cell = elementCreator('div', '-1');
 
-      const cellElement = handleCellFocus.findCellWithTabStop(rootElement);
+      const cellElement = handleAccessibility.findCellWithTabStop(rootElement);
 
       expect(cellElement).to.be.null;
     });
@@ -138,7 +138,7 @@ describe('handle-key-press', () => {
     };
 
     it('should call setFocusedCellCoord with adjusted index, and keyboard.focus', () => {
-      handleCellFocus.handleClickToFocusBody(cellData, rootElement, setFocusedCellCoord, keyboard);
+      handleAccessibility.handleClickToFocusBody(cellData, rootElement, setFocusedCellCoord, keyboard);
       expect(cell.setAttribute).have.been.calledOnceWith('tabIndex', '-1');
       expect(setFocusedCellCoord).to.have.been.calledOnceWith([1, 0]);
       expect(keyboard.focus).to.have.been.calledOnce;
@@ -149,7 +149,7 @@ describe('handle-key-press', () => {
     const columnIndex = 2;
 
     it('should indirectly call updateFocus, setFocusedCellCoord and keyboard.focus', () => {
-      handleCellFocus.handleClickToFocusHead(columnIndex, rootElement, setFocusedCellCoord, keyboard);
+      handleAccessibility.handleClickToFocusHead(columnIndex, rootElement, setFocusedCellCoord, keyboard);
       expect(cell.setAttribute).have.been.calledOnceWith('tabIndex', '-1');
       expect(setFocusedCellCoord).to.have.been.calledOnceWith([0, 2]);
       expect(keyboard.focus).to.have.been.calledOnce;
@@ -169,7 +169,7 @@ describe('handle-key-press', () => {
     });
 
     it('should set tabindex on the first cell and not focus', () => {
-      handleCellFocus.handleResetFocus({
+      handleAccessibility.handleResetFocus({
         focusedCellCoord,
         rootElement,
         shouldRefocus,
@@ -185,7 +185,7 @@ describe('handle-key-press', () => {
     it('should set tabindex on the first cell and focus when shouldRefocus is true', () => {
       shouldRefocus.current = true;
 
-      handleCellFocus.handleResetFocus({
+      handleAccessibility.handleResetFocus({
         focusedCellCoord,
         rootElement,
         shouldRefocus,
@@ -206,7 +206,7 @@ describe('handle-key-press', () => {
         querySelector: () => cell,
       };
 
-      handleCellFocus.handleResetFocus({
+      handleAccessibility.handleResetFocus({
         focusedCellCoord,
         rootElement,
         shouldRefocus,
@@ -220,73 +220,18 @@ describe('handle-key-press', () => {
     });
   });
 
-  describe('handleNavigateTop', () => {
-    let scrollTo;
-    let tableSectionRef;
-    let rowHeight;
-
-    beforeEach(() => {
-      rowHeight = 100;
-      scrollTo = sinon.spy();
-      tableSectionRef = { current: { scrollTo } };
-    });
-
-    it('should not do anything when ref is not setup yet', () => {
-      tableSectionRef.current = {};
-
-      handleCellFocus.handleNavigateTop({ tableSectionRef, focusedCellCoord, rootElement });
-      expect(scrollTo).to.not.have.been.called;
-    });
-
-    it('should scroll to the top when you reach the top two rows', () => {
-      focusedCellCoord = [1, 0];
-
-      handleCellFocus.handleNavigateTop({ tableSectionRef, focusedCellCoord, rootElement });
-      expect(scrollTo).to.have.been.calledOnce;
-    });
-
-    it('should scroll upwards automatically if it detects the cursor gets behind <TableHead />', () => {
-      const SCROLL_TOP_IDX = 7;
-      focusedCellCoord = [8, 0];
-      tableSectionRef = { current: { scrollTo, scrollTop: SCROLL_TOP_IDX * rowHeight } };
-      rootElement = {
-        getElementsByClassName: (query) => {
-          if (query === 'sn-table-head-cell') {
-            return [{ offsetHeight: 128 }];
-          }
-
-          return Array.from(Array(10).keys()).map((idx) => {
-            const rowCell = {
-              offsetHeight: rowHeight,
-              offsetTop: idx * rowHeight,
-            };
-
-            return { getElementsByClassName: () => [rowCell] };
-          });
-        },
-      };
-      // targetOffsetTop = tableSection.current.scrollTop - cell.offsetHeight - tableHead.offsetHeight;
-      // 700 - 100 - 128 = 472 => so our scrollTo function migth be called with 600
-      const targetOffsetTop = 472;
-
-      handleCellFocus.handleNavigateTop({ tableSectionRef, focusedCellCoord, rootElement });
-      expect(scrollTo).to.have.been.calledOnce;
-      expect(scrollTo).to.have.been.calledOnceWith({ top: targetOffsetTop, behavior: 'smooth' });
-    });
-  });
-
   describe('getCellSelectionStatusNote', () => {
     let rows;
 
     it('should return singular notation for array with single item', () => {
       rows = [1];
-      const result = handleCellFocus.getCellSelectionStatusNote(rows, translator);
+      const result = handleAccessibility.getCellSelectionStatusNote(rows, translator);
       expect(result).to.equal('SNTable.SelectionLabel.OneSelectedValue');
     });
 
     it('should return plural notation for arrays with multiple items', () => {
       rows = [1, 2, 3];
-      const result = handleCellFocus.getCellSelectionStatusNote(rows, translator);
+      const result = handleAccessibility.getCellSelectionStatusNote(rows, translator);
       expect(result).to.equal(`SNTable.SelectionLabel.SelectedValues`);
     });
   });
@@ -299,7 +244,7 @@ describe('handle-key-press', () => {
     beforeEach(() => {
       selectionState = { rows: [1, 2] };
       isActiveElementInTable = true;
-      getCellSrNotation = handleCellFocus.getMemoisedSrNotation();
+      getCellSrNotation = handleAccessibility.getMemoisedSrNotation();
     });
 
     it('should return empty string while we are in first row', () => {
@@ -347,7 +292,7 @@ describe('handle-key-press', () => {
     it('should return `value selected` and selected rows count while we are selecting multiple rows', () => {
       focusedCellCoord = [2, 1];
       selectionState = { rows: ['row#01', 'row#02'] };
-      getCellSrNotation = handleCellFocus.getMemoisedSrNotation(1);
+      getCellSrNotation = handleAccessibility.getMemoisedSrNotation(1);
 
       const notation = getCellSrNotation({
         focusedCellCoord,
@@ -363,7 +308,7 @@ describe('handle-key-press', () => {
     it('should be able to deselect previously selected value', () => {
       focusedCellCoord = [2, 1];
       selectionState = { rows: ['row#01', 'row#02'] };
-      getCellSrNotation = handleCellFocus.getMemoisedSrNotation(3);
+      getCellSrNotation = handleAccessibility.getMemoisedSrNotation(3);
 
       const notation = getCellSrNotation({
         focusedCellCoord,
@@ -379,7 +324,7 @@ describe('handle-key-press', () => {
     it('should be able to detect if cell has been selected while changing the focus to cell if we are in selection mode', () => {
       focusedCellCoord = [1, 1];
       selectionState = { rows: ['row#01', 'row#02'] };
-      getCellSrNotation = handleCellFocus.getMemoisedSrNotation(2);
+      getCellSrNotation = handleAccessibility.getMemoisedSrNotation(2);
       cell = global.document.createElement('td');
       cell.classList.add('selected');
       rootElement = {
@@ -400,7 +345,7 @@ describe('handle-key-press', () => {
     it('should be able to detect if cell has not been selected while changing the focus to cell if we are in selection mode', () => {
       focusedCellCoord = [2, 1];
       selectionState = { rows: ['row#01', 'row#02'] };
-      getCellSrNotation = handleCellFocus.getMemoisedSrNotation(2);
+      getCellSrNotation = handleAccessibility.getMemoisedSrNotation(2);
 
       const notation = getCellSrNotation({
         focusedCellCoord,
@@ -416,7 +361,7 @@ describe('handle-key-press', () => {
     it('should convey selection exited when we deselect very last selected cell in column', () => {
       focusedCellCoord = [2, 1];
       selectionState = { rows: [] };
-      getCellSrNotation = handleCellFocus.getMemoisedSrNotation(1);
+      getCellSrNotation = handleAccessibility.getMemoisedSrNotation(1);
 
       const notation = getCellSrNotation({
         focusedCellCoord,
@@ -448,21 +393,21 @@ describe('handle-key-press', () => {
     });
 
     it('should call blur when currentTarget doesnt contain relatedTarget and shouldRefocus is false', () => {
-      handleCellFocus.handleFocusoutEvent(evt, shouldRefocus, blur);
+      handleAccessibility.handleFocusoutEvent(evt, shouldRefocus, blur);
       expect(blur).to.have.been.calledOnceWith(false);
     });
 
     it('should not call blur when currentTarget contains relatedTarget', () => {
       containsRelatedTarget = true;
 
-      handleCellFocus.handleFocusoutEvent(evt, shouldRefocus, blur);
+      handleAccessibility.handleFocusoutEvent(evt, shouldRefocus, blur);
       expect(blur).to.not.have.been.called;
     });
 
     it('should not call blur when shouldRefocus is true', () => {
       shouldRefocus.current = true;
 
-      handleCellFocus.handleFocusoutEvent(evt, shouldRefocus, blur);
+      handleAccessibility.handleFocusoutEvent(evt, shouldRefocus, blur);
       expect(blur).to.not.have.been.called;
     });
   });
