+++ src/table/utils/handle-accessibility.js
@@ -63,32 +63,6 @@ export const handleResetFocus = ({
   setFocusedCellCoord(nextCell);
 };
 
-export const handleNavigateTop = ({ tableSectionRef, focusedCellCoord, rootElement }) => {
-  const MIN_ROW_COUNT = 2;
-
-  if (!tableSectionRef.current?.scrollTo) return;
-
-  if (focusedCellCoord[0] < MIN_ROW_COUNT) {
-    tableSectionRef.current.scrollTo({
-      top: 0,
-      behavior: 'smooth',
-    });
-  } else {
-    const [x, y] = focusedCellCoord;
-    const tableHead = rootElement.getElementsByClassName('sn-table-head-cell')[0];
-    const rowElements = rootElement.getElementsByClassName('sn-table-row');
-    const cell = rowElements[x]?.getElementsByClassName('sn-table-cell')[y];
-
-    if (cell.offsetTop - tableHead.offsetHeight - cell.offsetHeight <= tableSectionRef.current.scrollTop) {
-      const targetOffsetTop = tableSectionRef.current.scrollTop - cell.offsetHeight - tableHead.offsetHeight;
-      tableSectionRef.current.scrollTo({
-        top: Math.max(0, targetOffsetTop),
-        behavior: 'smooth',
-      });
-    }
-  }
-};
-
 export const getCellSelectionStatusNote = (rows, translator) =>
   rows.length === 1
     ? translator.get('SNTable.SelectionLabel.OneSelectedValue')
