+++ src/table/utils/handle-scroll.js
@@ -1,4 +1,4 @@
-const handleScroll = (evt, tableSectionRef) => {
+export const handleScroll = (evt, tableSectionRef) => {
   evt.stopPropagation();
   let { scrollLeft } = tableSectionRef.current;
   const newScrollLeft = scrollLeft + evt.deltaX;
@@ -9,4 +9,28 @@ const handleScroll = (evt, tableSectionRef) => {
   }
 };
 
-export default handleScroll;
+export const handleNavigateTop = ({ tableSectionRef, focusedCellCoord, rootElement }) => {
+  const MIN_ROW_COUNT = 2;
+
+  if (!tableSectionRef.current?.scrollTo) return;
+
+  if (focusedCellCoord[0] < MIN_ROW_COUNT) {
+    tableSectionRef.current.scrollTo({
+      top: 0,
+      behavior: 'smooth',
+    });
+  } else {
+    const [x, y] = focusedCellCoord;
+    const tableHead = rootElement.getElementsByClassName('sn-table-head-cell')[0];
+    const rowElements = rootElement.getElementsByClassName('sn-table-row');
+    const cell = rowElements[x]?.getElementsByClassName('sn-table-cell')[y];
+
+    if (cell.offsetTop - tableHead.offsetHeight - cell.offsetHeight <= tableSectionRef.current.scrollTop) {
+      const targetOffsetTop = tableSectionRef.current.scrollTop - cell.offsetHeight - tableHead.offsetHeight;
+      tableSectionRef.current.scrollTo({
+        top: Math.max(0, targetOffsetTop),
+        behavior: 'smooth',
+      });
+    }
+  }
+};
